Code changes to get unit tests to compile with Real of Fixed32.

+++ UnitTests/UnitVec2.cpp
@@ -77,26 +77,26 @@ TEST(UnitVec2, BottomIsFwdPerpOfRight)
 
 TEST(UnitVec2, ByAngleInDegreesNearOriented)
 {
-    EXPECT_NEAR(GetX(UnitVec2(Angle(0))), GetX(UnitVec2::GetRight()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2(Angle(0))), GetY(UnitVec2::GetRight()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2(Angle{Real(90) * Degree})), GetX(UnitVec2::GetTop()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2(Angle{Real(90) * Degree})), GetY(UnitVec2::GetTop()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2(Angle{Real(180) * Degree})), GetX(UnitVec2::GetLeft()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2(Angle{Real(180) * Degree})), GetY(UnitVec2::GetLeft()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2(Angle{Real(270) * Degree})), GetX(UnitVec2::GetBottom()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2(Angle{Real(270) * Degree})), GetY(UnitVec2::GetBottom()), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2(Angle(       0          )))), static_cast<double>(GetX(UnitVec2::GetRight())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2(Angle(       0          )))), static_cast<double>(GetY(UnitVec2::GetRight())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2(Angle{Real( 90) * Degree}))), static_cast<double>(GetX(UnitVec2::GetTop())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2(Angle{Real( 90) * Degree}))), static_cast<double>(GetY(UnitVec2::GetTop())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2(Angle{Real(180) * Degree}))), static_cast<double>(GetX(UnitVec2::GetLeft())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2(Angle{Real(180) * Degree}))), static_cast<double>(GetY(UnitVec2::GetLeft())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2(Angle{Real(270) * Degree}))), static_cast<double>(GetX(UnitVec2::GetBottom())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2(Angle{Real(270) * Degree}))), static_cast<double>(GetY(UnitVec2::GetBottom())), 0.0001);
 }
 
 TEST(UnitVec2, ByAngleInRadiansNearOriented)
 {
-    EXPECT_NEAR(GetX(UnitVec2((Pi * Real(0) / Real(2)) * Radian)), GetX(UnitVec2::GetRight()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2((Pi * Real(0) / Real(2)) * Radian)), GetY(UnitVec2::GetRight()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2((Pi * Real(1) / Real(2)) * Radian)), GetX(UnitVec2::GetTop()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2((Pi * Real(1) / Real(2)) * Radian)), GetY(UnitVec2::GetTop()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2((Pi * Real(2) / Real(2)) * Radian)), GetX(UnitVec2::GetLeft()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2((Pi * Real(2) / Real(2)) * Radian)), GetY(UnitVec2::GetLeft()), 0.0001);
-    EXPECT_NEAR(GetX(UnitVec2((Pi * Real(3) / Real(2)) * Radian)), GetX(UnitVec2::GetBottom()), 0.0001);
-    EXPECT_NEAR(GetY(UnitVec2((Pi * Real(3) / Real(2)) * Radian)), GetY(UnitVec2::GetBottom()), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2((Pi * Real(0) / Real(2)) * Radian))), static_cast<double>(GetX(UnitVec2::GetRight())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2((Pi * Real(0) / Real(2)) * Radian))), static_cast<double>(GetY(UnitVec2::GetRight())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2((Pi * Real(1) / Real(2)) * Radian))), static_cast<double>(GetX(UnitVec2::GetTop())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2((Pi * Real(1) / Real(2)) * Radian))), static_cast<double>(GetY(UnitVec2::GetTop())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2((Pi * Real(2) / Real(2)) * Radian))), static_cast<double>(GetX(UnitVec2::GetLeft())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2((Pi * Real(2) / Real(2)) * Radian))), static_cast<double>(GetY(UnitVec2::GetLeft())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetX(UnitVec2((Pi * Real(3) / Real(2)) * Radian))), static_cast<double>(GetX(UnitVec2::GetBottom())), 0.0001);
+    EXPECT_NEAR(static_cast<double>(GetY(UnitVec2((Pi * Real(3) / Real(2)) * Radian))), static_cast<double>(GetY(UnitVec2::GetBottom())), 0.0001);
 }
 
 TEST(UnitVec2, GetForInvalid)
